requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
bcrypt==4.1.2
pyotp==2.9.0
qrcode[pil]==7.4.2
requests==2.31.0
python-dotenv==1.0.0
email_validator==2.1.0.post1
